update check ota
